requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
spotipy==2.24.0
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
python-dotenv==1.0.1
httpx==0.27.2

# PostgreSQL
sqlalchemy==2.0.36
asyncpg==0.31.0
psycopg2-binary==2.9.11
greenlet>=3.0.0
